update(event-emitter): onの調査
